Sandfly 5.2.0 release

+++ start_scripts/start_server.sh
@@ -110,6 +110,17 @@ if [ $? -ne 0 ]; then
     exit 1
 fi
 
+# See if we can run Docker
+which docker >/dev/null 2>&1 || { echo "Unable to locate docker binary; please install Docker."; exit 1; }
+docker version >/dev/null 2>&1 || { echo "This script must be run as root or as a user with access to the Docker daemon."; exit 1; }
+
+# Load images if offline bundle is present and not already loaded
+../setup/setup_scripts/load_images.sh
+if [ "$?" -ne 0 ]; then
+  echo "Error loading container images."
+  exit 1
+fi
+
 # Old versions of Sandfly may have left behind a temporary volume for the
 # old rabbit container. Clean it up if present.
 docker volume rm sandfly-rabbitmq-tmp-vol 2>/dev/null
